+++ libvast/src/qualified_record_field.cpp
@@ -11,46 +11,34 @@
  * contained in the LICENSE file.                                             *
  ******************************************************************************/
 
-#pragma once
+#include "vast/qualified_record_field.hpp"
 
-#include <iterator>
-#include <vector>
-
-#include "vast/concept/printable/core/printer.hpp"
-#include "vast/concept/support/detail/attr_fold.hpp"
+#include "vast/concept/hashable/uhash.hpp"
+#include "vast/concept/hashable/xxhash.hpp"
 
 namespace vast {
 
-template <class Lhs, class Rhs>
-class list_printer : public printer<list_printer<Lhs, Rhs>> {
-public:
-  using lhs_attribute = typename Lhs::attribute;
-  using rhs_attribute = typename Rhs::attribute;
-  using attribute = detail::attr_fold_t<std::vector<lhs_attribute>>;
+bool operator==(const qualified_record_field& x,
+                const qualified_record_field& y) {
+  return x.fqn == y.fqn && x.type == y.type;
+}
 
-  list_printer(Lhs lhs, Rhs rhs) : lhs_{std::move(lhs)}, rhs_{std::move(rhs)} {
-    // nop
-  }
+bool operator<(const qualified_record_field& x,
+               const qualified_record_field& y) {
+  return std::tie(x.fqn, x.type) < std::tie(y.fqn, y.type);
+}
 
-  template <class Iterator, class Attribute>
-  bool print(Iterator& out, const Attribute& a) const {
-    using std::begin;
-    using std::end;
-    auto f = begin(a);
-    auto l = end(a);
-    if (f == l)
-      return true;
-    if (!lhs_.print(out, *f))
-      return false;
-    for (++f; f != l; ++f)
-      if (!(rhs_.print(out, unused) && lhs_.print(out, *f)))
-        return false;
-    return true;
-  }
-
-private:
-  Lhs lhs_;
-  Rhs rhs_;
-};
+record_field as_record_field(const qualified_record_field& qf) {
+  return {qf.fqn, qf.type};
+}
 
 } // namespace vast
+
+namespace std {
+
+size_t hash<vast::qualified_record_field>::operator()(
+  const vast::qualified_record_field& x) const {
+  return vast::uhash<vast::xxhash64>{}(x);
+}
+
+} // namespace std