+++ libvast/src/system/index.cpp
@@ -34,7 +34,7 @@
 #include "vast/save.hpp"
 #include "vast/system/accountant.hpp"
 #include "vast/system/evaluator.hpp"
-#include "vast/system/index.hpp"
+#include "vast/system/index_common.hpp"
 #include "vast/system/partition.hpp"
 #include "vast/system/query_supervisor.hpp"
 #include "vast/system/spawn_indexer.hpp"
@@ -42,13 +42,11 @@
 #include "vast/table_slice.hpp"
 
 #include <caf/all.hpp>
-#include <caf/detail/unordered_flat_map.hpp>
 
 #include <chrono>
 #include <deque>
 #include <unordered_set>
 
-using namespace caf;
 using namespace std::chrono;
 
 namespace vast::system {
@@ -87,6 +85,10 @@ index_state::index_state(caf::stateful_actor<index_state>* self)
 
 index_state::~index_state() {
   VAST_VERBOSE(self, "tearing down");
+  if (active != nullptr) {
+    [[maybe_unused]] auto unregistered = stage->out().unregister(active.get());
+    VAST_ASSERT(unregistered);
+  }
   flush_to_disk();
 }
 
@@ -103,7 +105,7 @@ caf::error index_state::init(const path& dir, size_t max_partition_size,
   this->taste_partitions = taste_partitions;
   if (auto a = self->system().registry().get(accountant_atom::value)) {
     namespace defs = defaults::system;
-    this->accountant = actor_cast<accountant_type>(a);
+    this->accountant = caf::actor_cast<accountant_type>(a);
     self->send(this->accountant, announce_atom::value, "index");
     self->delayed_send(self, defs::telemetry_rate, telemetry_atom::value);
   }
@@ -167,7 +169,7 @@ caf::error index_state::flush_to_disk() {
       if (auto err = active->flush_to_disk())
         return err;
     // Flush all unpersisted partitions. This only writes the meta state of
-    // each table_indexer. For actually writing the contents of each INDEXER we
+    // each partition. For actually writing the contents of each INDEXER we
     // need to rely on messaging.
     for (auto& kvp : unpersisted)
       if (auto err = kvp.first->flush_to_disk())
@@ -237,27 +239,18 @@ void index_state::send_report() {
   auto min_rate = std::numeric_limits<double>::infinity();
   auto max_rate = -std::numeric_limits<double>::infinity();
   auto append_report = [&](partition& p) {
-    for (auto& [layout, ti] : p.table_indexers_) {
-      for (size_t i = 0; i < ti.measurements_.size(); ++i) {
-#ifdef VAST_MEASUREMENT_MUTEX_WORKAROUND
-        ti.measurements_[i].mutex.lock();
-        auto tmp = static_cast<measurement>(ti.measurements_[i]);
-        ti.measurements_[i].reset();
-        ti.measurements_[i].mutex.unlock();
-#else
-        auto tmp = std::atomic_exchange(&(ti.measurements_[i]), measurement{});
-#endif
-        if (tmp.events > 0) {
-          r.push_back({layout.name() + "." + layout.fields[i].name, tmp});
-          double rate = tmp.events * 1'000'000'000.0 / tmp.duration.count();
-          if (rate < min_rate) {
-            min_rate = rate;
-            min = tmp;
-          }
-          if (rate > max_rate) {
-            max_rate = rate;
-            max = tmp;
-          }
+    for (size_t i = 0; i < p.measurements_.size(); ++i) {
+      auto tmp = collect(p.measurements_[i]);
+      if (tmp.events > 0) {
+        r.push_back({as_vector(p.indexers_)[i].first.fqn, tmp});
+        double rate = tmp.events * 1'000'000'000.0 / tmp.duration.count();
+        if (rate < min_rate) {
+          min_rate = rate;
+          min = tmp;
+        }
+        if (rate > max_rate) {
+          max_rate = rate;
+          max = tmp;
         }
       }
     }
@@ -286,6 +279,8 @@ void index_state::reset_active_partition() {
   // Persist meta data and the state of all INDEXER actors when the active
   // partition gets replaced becomes full.
   if (active != nullptr) {
+    [[maybe_unused]] auto unregistered = stage->out().unregister(active.get());
+    VAST_ASSERT(unregistered);
     if (auto err = active->flush_to_disk())
       VAST_ERROR(self, "failed to persist active partition");
     // Store this partition as unpersisted to make sure we're not attempting
@@ -300,7 +295,14 @@ void index_state::reset_active_partition() {
   if (auto err = flush_statistics())
     VAST_ERROR(self, "failed to persist the statistics");
   active = make_partition();
+  stage->out().register_partition(active.get());
   active_partition_indexers = 0;
+}
+
+partition* index_state::get_or_add_partition(const table_slice_ptr& slice) {
+  if (!active || active->capacity() < slice->rows())
+    reset_active_partition();
+  return active.get();
 }
 
 partition_ptr index_state::make_partition() {
@@ -312,14 +314,14 @@ partition_ptr index_state::make_partition(uuid id) {
   return std::make_unique<partition>(this, std::move(id), max_partition_size);
 }
 
-caf::actor index_state::make_indexer(path dir, type column_type, size_t column,
+caf::actor index_state::make_indexer(path filename, type column_type,
                                      uuid partition_id, atomic_measurement* m) {
-  VAST_TRACE(VAST_ARG(dir), VAST_ARG(column_type), VAST_ARG(column),
-             VAST_ARG(index), VAST_ARG(partition_id));
+  VAST_TRACE(VAST_ARG(dir), VAST_ARG(column_type), VAST_ARG(index),
+             VAST_ARG(partition_id));
   caf::settings index_opts;
   index_opts["cardinality"] = max_partition_size;
-  return factory(self, std::move(dir), std::move(column_type),
-                 std::move(index_opts), column, self, partition_id, m);
+  return factory(self, std::move(filename), std::move(column_type),
+                 std::move(index_opts), self, partition_id, m);
 }
 
 void index_state::decrement_indexer_count(uuid partition_id) {
@@ -346,9 +348,7 @@ partition* index_state::find_unpersisted(const uuid& id) {
   return i != unpersisted.end() ? i->first.get() : nullptr;
 }
 
-using pending_query_map = caf::detail::unordered_flat_map<uuid, evaluation_map>;
-
-pending_query_map
+index_state::pending_query_map
 index_state::build_query_map(lookup_state& lookup, uint32_t num_partitions) {
   VAST_TRACE(VAST_ARG(lookup), VAST_ARG(num_partitions));
   if (num_partitions == 0 || lookup.partitions.empty())
@@ -418,9 +418,9 @@ void index_state::notify_flush_listeners() {
   flush_listeners.clear();
 }
 
-behavior index(stateful_actor<index_state>* self, const path& dir,
-               size_t max_partition_size, size_t in_mem_partitions,
-               size_t taste_partitions, size_t num_workers) {
+caf::behavior index(caf::stateful_actor<index_state>* self, const path& dir,
+                    size_t max_partition_size, size_t in_mem_partitions,
+                    size_t taste_partitions, size_t num_workers) {
   VAST_TRACE(VAST_ARG(dir), VAST_ARG(max_partition_size),
              VAST_ARG(in_mem_partitions), VAST_ARG(taste_partitions),
              VAST_ARG(num_workers));
@@ -433,7 +433,7 @@ behavior index(stateful_actor<index_state>* self, const path& dir,
     self->quit(std::move(err));
     return {};
   }
-  self->set_exit_handler([=](const exit_msg& msg) {
+  self->set_exit_handler([=](const caf::exit_msg& msg) {
     VAST_DEBUG(self, "received exit from", msg.source,
                "with reason:", msg.reason);
     self->state.send_report();
@@ -455,7 +455,7 @@ behavior index(stateful_actor<index_state>* self, const path& dir,
       // Sanity check.
       if (self->current_sender() == nullptr) {
         VAST_ERROR(self, "got an anonymous query (ignored)");
-        respond(sec::invalid_argument);
+        respond(caf::sec::invalid_argument);
         return;
       }
       auto& st = self->state;
@@ -518,7 +518,7 @@ behavior index(stateful_actor<index_state>* self, const path& dir,
         VAST_ERROR(self, "got an anonymous query (ignored)");
         return;
       }
-      auto client = actor_cast<actor>(self->current_sender());
+      auto client = caf::actor_cast<caf::actor>(self->current_sender());
       auto iter = st.pending.find(query_id);
       if (iter == st.pending.end()) {
         VAST_WARNING(self, "got a request for unknown query ID", query_id);
@@ -562,13 +562,13 @@ behavior index(stateful_actor<index_state>* self, const path& dir,
       namespace defs = defaults::system;
       self->delayed_send(self, defs::telemetry_rate, telemetry_atom::value);
     },
-    [=](subscribe_atom, flush_atom, actor& listener) {
+    [=](subscribe_atom, flush_atom, caf::actor& listener) {
       self->state.add_flush_listener(std::move(listener));
     });
   return {[=](worker_atom, caf::actor& worker) {
             auto& st = self->state;
             st.idle_workers.emplace_back(std::move(worker));
-            self->become(keep_behavior, st.has_worker);
+            self->become(caf::keep_behavior, st.has_worker);
           },
           [=](done_atom, uuid partition_id) {
             self->state.decrement_indexer_count(partition_id);
@@ -586,7 +586,7 @@ behavior index(stateful_actor<index_state>* self, const path& dir,
             self->delayed_send(self, defs::telemetry_rate,
                                telemetry_atom::value);
           },
-          [=](subscribe_atom, flush_atom, actor& listener) {
+          [=](subscribe_atom, flush_atom, caf::actor& listener) {
             self->state.add_flush_listener(std::move(listener));
           }};
 }