+++ libvast/src/system/indexer_stage_driver.cpp
@@ -13,23 +13,21 @@
 
 #include "vast/system/indexer_stage_driver.hpp"
 
-#include <caf/downstream.hpp>
-#include <caf/event_based_actor.hpp>
-#include <caf/scheduled_actor.hpp>
-#include <caf/stream_manager.hpp>
-
 #include "vast/logger.hpp"
 #include "vast/meta_index.hpp"
 #include "vast/system/index.hpp"
 #include "vast/system/partition.hpp"
 #include "vast/table_slice.hpp"
 
-namespace vast::system {
+#include <caf/downstream.hpp>
+#include <caf/event_based_actor.hpp>
+#include <caf/outbound_path.hpp>
+#include <caf/scheduled_actor.hpp>
+#include <caf/stream_manager.hpp>
 
-bool indexer_stage_selector::operator()(const indexer_stage_filter& f,
-                                        const table_slice_ptr& x) const {
-  return f == x->layout();
-}
+#include <cstddef>
+
+namespace vast::system {
 
 indexer_stage_driver::indexer_stage_driver(downstream_manager_type& dm,
                                            self_pointer self)
@@ -41,61 +39,16 @@ indexer_stage_driver::~indexer_stage_driver() noexcept {
   // nop
 }
 
-void indexer_stage_driver::process(downstream_type& out, batch_type& slices) {
+void indexer_stage_driver::process(downstream_type&, batch_type& slices) {
   VAST_TRACE(CAF_ARG(slices));
   VAST_ASSERT(!slices.empty());
   auto& st = self_->state;
   for (auto& slice : slices) {
-    // Spin up an initial partition if needed.
-    if (st.active == nullptr)
-      st.reset_active_partition();
-    // Update meta index.
-    st.meta_idx.add(st.active->id(), *slice);
-    // Update statistics.
     auto& layout = slice->layout();
     st.stats.layouts[layout.name()].count += slice->rows();
-    // Start new INDEXER actors when needed and add it to the stream.
-    if (auto ti = st.active->get_or_add(layout)) {
-      auto [meta_x, added] = *ti;
-      if (added) {
-        VAST_DEBUG(st.self, "added a new table_indexer for layout", layout);
-        if (auto err = meta_x.init()) {
-          VAST_ERROR(st.self, "failed to initialize table_indexer for layout",
-                     layout, "-> all incoming logs get dropped!");
-        } else {
-          meta_x.spawn_indexers();
-          for (auto& x : meta_x.indexers()) {
-            // We'll have invalid handles at all fields with skip attribute.
-            if (x) {
-              auto slt = out_.parent()
-                           ->add_unchecked_outbound_path<output_type>(x);
-              VAST_DEBUG(st.self, "spawned new INDEXER at slot", slt);
-              out_.set_filter(slt, layout);
-              st.active_partition_indexers++;
-            }
-          }
-        }
-      }
-      // Add all rows IDs to the meta indexer.
-      meta_x.add(slice);
-    }
-    // Ship event to the INDEXER actors.
-    auto slice_size = slice->rows();
-    out.push(std::move(slice));
-    // Reset the manager and all outbound paths when finalizing a partition.
-    if (st.active->capacity() <= slice_size) {
-      VAST_DEBUG(st.self, "closes slots on full partition",
-                 out_.open_path_slots());
-      VAST_ASSERT(out_.buf().size() != 0);
-      out_.fan_out_flush();
-      VAST_ASSERT(out_.buf().size() == 0);
-      out_.force_emit_batches();
-      out_.close();
-      st.reset_active_partition();
-      VAST_ASSERT(st.active->layouts().empty());
-    } else {
-      st.active->reduce_capacity(slice_size);
-    }
+    auto part = st.get_or_add_partition(slice);
+    st.meta_idx.add(part->id(), *slice);
+    part->add(std::move(slice));
   }
 }