+++ libvast/vast/arrow_table_slice.hpp
@@ -67,8 +67,7 @@ public:
 
   caf::error deserialize(caf::deserializer& source) override;
 
-  void
-  append_column_to_index(size_type col, vast::value_index& idx) const override;
+  void append_column_to_index(size_type col, value_index& idx) const override;
 
   caf::atom_value implementation_id() const noexcept override;