+++ libvast/vast/system/fwd.hpp
@@ -28,23 +28,19 @@ class configuration;
 class default_application;
 class export_command;
 class import_command;
-class indexer_stage_driver;
 class node_command;
-class partition;
 class pcap_reader_command;
 class pcap_writer_command;
 class remote_command;
 class sink_command;
 class source_command;
 class start_command;
-class table_indexer;
 
 // -- structs ------------------------------------------------------------------
 
 #ifdef VAST_MEASUREMENT_MUTEX_WORKAROUND
 struct atomic_measurement;
 #endif
-struct index_state;
 struct measurement;
 struct node_state;
 struct query_status;
@@ -56,8 +52,6 @@ struct spawn_arguments;
 using atomic_measurement = std::atomic<measurement>;
 #endif
 using node_actor = caf::stateful_actor<node_state>;
-using partition_ptr = std::unique_ptr<partition>;
-using table_indexer_ptr = caf::intrusive_ptr<table_indexer>;
 
 
 } // namespace vast::system