+++ libvast_test/src/dummy_index.cpp
@@ -23,7 +23,6 @@ using atomic_measurement = vast::system::atomic_measurement;
 
 CAF_ALLOW_UNSAFE_MESSAGE_TYPE(void_fun)
 
-using namespace caf;
 using namespace vast;
 
 using vast::system::index_state;
@@ -32,19 +31,21 @@ namespace fixtures {
 
 namespace {
 
-behavior dummy_indexer(stateful_actor<dummy_index::dummy_indexer_state>*) {
-  return {[](ok_atom) {
+caf::behavior
+dummy_indexer(caf::stateful_actor<dummy_index::dummy_indexer_state>*) {
+  return {[](caf::ok_atom) {
     // nop
   }};
 }
 
-actor spawn_dummy_indexer(local_actor* self, path, type, caf::settings, size_t,
-                          caf::actor, uuid, atomic_measurement*) {
+caf::actor
+spawn_dummy_indexer(caf::local_actor* self, path, type, caf::settings,
+                    caf::actor, uuid, atomic_measurement*) {
   return self->spawn(dummy_indexer);
 }
 
-behavior dummy_index_actor(stateful_actor<index_state>* self,
-                           path dir) {
+caf::behavior
+dummy_index_actor(caf::stateful_actor<index_state>* self, path dir) {
   self->state.init(std::move(dir), std::numeric_limits<size_t>::max(), 10, 5);
   self->state.factory = spawn_dummy_indexer;
   return {[](std::function<void()> f) { f(); }};
@@ -55,21 +56,19 @@ behavior dummy_index_actor(stateful_actor<index_state>* self,
 dummy_index::dummy_index() {
   idx_handle = sys.spawn(dummy_index_actor, directory);
   run();
-  idx_state = &deref<stateful_actor<index_state>>(idx_handle).state;
+  idx_state = &deref<caf::stateful_actor<index_state>>(idx_handle).state;
 }
 
 dummy_index::~dummy_index() {
-  anon_send_exit(idx_handle, exit_reason::kill);
+  anon_send_exit(idx_handle, caf::exit_reason::kill);
   run();
 }
 
 void dummy_index::run_in_index(std::function<void()> f) {
-  mailbox_element_vals<void_fun> tmp{nullptr,
-                                     make_message_id(),
-                                     {},
-                                     std::move(f)};
+  caf::mailbox_element_vals<void_fun> tmp{
+    nullptr, caf::make_message_id(), {}, std::move(f)};
   auto ctx = sys.dummy_execution_unit();
-  deref<stateful_actor<index_state>>(idx_handle).activate(ctx, tmp);
+  deref<caf::stateful_actor<index_state>>(idx_handle).activate(ctx, tmp);
 }
 
 } // namespace fixtures