+++ CHANGELOG.md
@@ -14,6 +14,11 @@ Every entry has a category for which we use the following visual abbreviations:
   script located at `scripts/vast-completions.bash` in the VAST source tree.
   [#833](https://github.com/tenzir/vast/pull/833)
 
+- 🔄 The index specific options `max-partition-size`, `max-resident-partitions`,
+  `max-taste-partitions`, and `max-queries` can now be specified on the command
+  line when starting a node.
+  [#728](https://github.com/tenzir/vast/pull/728)
+
 - 🎁 Packet drop statistics are now reported to the accountant for PCAP import,
   and are available using the keys `pcap-reader.recv`, `pcap-reader.drop`, and
   `pcap-reader.ifdrop` in the `vast.statistics` event. If the number of dropped